LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-rowspan-003-expected.txt

Visibility collapse

Spec
Setting a row to visibility:collapse changes table height but not width. Here the spanning cell and another cell are in the same row.

aaaa
bbbb
spanning	cccc
dddd
eeee
Bottom table is identical to top except third row has been collapsed.
aaaa
bbbb
dddd
eeee

PASS spanning row visibility:collapse doesn't change table width
FAIL collapsed row has zero height assert_equals: collapsed row has zero height expected 0 but got 29
PASS first row height doesn't change
PASS second row height doesn't change
PASS fourth row height doesn't change
PASS fifth row height doesn't change
